samples/ml/ml_jobs/single-node/pytorch-cifar10/requirements.txt
snowflake-ml-python>=1.7.4